Solves the route, but logging output is tough to read.

+++ src/main/java/models/SearchNode.java
@@ -6,9 +6,9 @@ import java.util.Objects;
 public class SearchNode {
     private final byte x;
     private final byte y;
-    private final byte depth;
+    private final int depth;
 
-    public SearchNode(byte x, byte y, byte depth) {
+    public SearchNode(byte x, byte y, int depth) {
         if (x < 0) {
             throw new IllegalArgumentException(String.format("Value supplied as x must not be negative. %d was supplied.", x));
         }
@@ -34,7 +34,7 @@ public class SearchNode {
         return y;
     }
 
-    public byte getDepth() {
+    public int getDepth() {
         return depth;
     }
 
@@ -47,19 +47,19 @@ public class SearchNode {
         ArrayList<SearchNode> neighbours = new ArrayList<SearchNode>();
 
         if (canMoveLeft) {
-            neighbours.add(new SearchNode((byte)(getX() - 1), getY(), (byte)(getDepth() + 1)));
+            neighbours.add(new SearchNode((byte)(getX() - 1), getY(), getDepth() + 1));
         }
 
         if (canMoveUp) {
-            neighbours.add(new SearchNode(getX(), (byte)(getY() - 1), (byte)(getDepth() + 1)));
+            neighbours.add(new SearchNode(getX(), (byte)(getY() - 1), getDepth() + 1));
         }
 
         if (canMoveRight) {
-            neighbours.add(new SearchNode((byte)(getX() + 1), getY(), (byte)(getDepth() + 1)));
+            neighbours.add(new SearchNode((byte)(getX() + 1), getY(), getDepth() + 1));
         }
 
         if (canMoveDown) {
-            neighbours.add(new SearchNode(getX(), (byte)(getY() + 1), (byte)(getDepth() + 1)));
+            neighbours.add(new SearchNode(getX(), (byte)(getY() + 1), getDepth() + 1));
         }
 
         return neighbours;
